refactor: extract chained-intent / multi-entity detection into _ChainMixin
Fourth and largest decomposition cut (~950 lines). Moves the contiguous
chained-intent + multi-entity block — _chained_intent_guidance,
_is_strong_chain_connector, _soft_connector_footer, _looks_like_slashed_compound,
_split_multi_entity, _path_matches_topic_loosely, _multi_entity_chain_guidance,
_half_resolves_to_top, _is_substantive_topic, _is_topic_shaped plus their
connector / operation-prefix regex tables and conjunction-affix constants —
into openzim_mcp/chain_detection.py as _ChainMixin.

External coupling is just zim_operations + _recase_from_original (declared
under TYPE_CHECKING); IntentParser is imported. All connector/operation
constants are cluster-local. Pure code movement, no behavior change.

Updated test_slashed_compound_helper_is_canonical_source to pin the new
single-definition site (chain_detection.py); the "exactly one declaration"
invariant it guards is unchanged.

simple_tools.py 5,194 → 4,247 (5,896 → 4,247 across the four mixin cuts;
-1,649 lines / -28%). Full suite green (2623 passed); mypy/flake8/isort clean.
